Projects/Beginner_Level_Agents_Projects/Travel_Planner_Agents/requirements.txt
pydantic
typing
google.generativeai
dotenv